// packages/core/src/nodeLayer/types/binding.ts
export type ModuleType =
  | 'ts'
  | 'js'
  | 'jsx'
  | 'tsx'
  | 'css'
  | 'html'
  | 'asset'
  | string;
// ...
export interface PluginTransformHookParam {
  moduleId: string;
  /// source content after load or transformed result of previous plugin
  content: string;
  /// module type after load
  moduleType: ModuleType;
  resolvedPath: string;
  query: [string, string][];
  meta: Record<string, string> | null;
  sourceMapChain: string[];
}
// ...
export interface OutputConfig {
  /**
   * Configure the file name of the output files which contain the entry modules. Prior to filename
   */
  entryFilename?: string;
  /**
   * Configure the name of all the output files
   */
  filename?: string;
  /**
   * Output dir that production files are emitted to.
   */
  path?: string;
  /**
   * resource loading prefix. for example, if publicPath is `https://xxx.cdn.comm`,
   * then the url output files in html will be `https://xxx.cdn.com/index_ecad.xxxx.js`
   *
   * default by `output.targetEnv`, if node, publicPath is `./`, if browser, publicPath is `/`
   */
  publicPath?: string;
  /**
   * the same as `filename`, but only for static assets like `.png`, `.jpg`
   */
  assetsFilename?: string;
  /**
   * Target execution environment of production files, browser or node. browser is equal to `browser-es2017`, node is equal to `node16`.
   * You can also set target env version like `node16`, `node-legacy`, 'browser-legacy`, 'browser-es2015', 'browser-2017', 'browser-esnext'. Farm will automatically downgrade syntax and inject polyfill according to the specified target env.
   * @default 'browser'
   */
  targetEnv?:
    | 'browser'
    | 'node'
    | 'node16'
    | 'node-legacy'
    | 'node-next'
    | 'browser-legacy'
    | 'browser-es2015'
    | 'browser-es2017'
    | 'browser-esnext';
  /**
   * output module format
   */
  format?: 'cjs' | 'esm';
}